refactor: Seperated files and fix for buildings heigth alignment with terrain
